Completed styling of the technologies section
