--- src/pages/technology/Tech.json
@@ -1,10 +1,0 @@
-[
-   {
-      "title": "AI",
-      "img_url": "https://upload.wikimedia.org/wikipedia/en/c/c8/Very_Black_screen.jpg"
-   },
-   {
-      "title": "Blockchain",
-      "img_url": "https://upload.wikimedia.org/wikipedia/commons/b/bc/IKB_191.jpg"
-   }
-]
