+++ src/sections/technology/Technology.module.scss
@@ -1,0 +1,18 @@
+.container {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    justify-content: center;
+    @media (min-width: 1100px) {
+        margin: 100px
+    }
+}
+
+.heading {
+    margin-top: 75px;
+    font-size: 40px;
+    @media (min-width: 650px) {
+        font-size: 60px;
+    }
+    
+}
